backend unit tests
