+++ back/.gitignore
@@ -4,3 +4,4 @@ node_modules
 *.log
 config
 *.log.gz
+coverage/
