+++ back/package.json
@@ -5,7 +5,8 @@
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "eslint": "node node_modules/eslint/bin/eslint.js ./"
+    "eslint": "node node_modules/eslint/bin/eslint.js ./",
+    "test": "BASKET_MODE=dev jest --forceExit --coverage"
   },
   "author": "Zubkov Konstantin",
   "license": "MIT",
@@ -25,6 +26,9 @@
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
-    "babel-preset-env": "^1.7.0"
+    "babel-preset-env": "^1.7.0",
+    "jest": "^27.5.1",
+    "jest-environment-node": "^27.5.1",
+    "supertest": "^6.3.4"
   }
 }
