+++ back/server.js
@@ -1,60 +1,14 @@
-const bodyParser = require('body-parser');
-const cookieParser = require('cookie-parser');
-const express = require('express');
-const uuid = require('uuid');
-
-const apiRoutes = require('./api/routes');
-const auth = require('./utils/auth');
+const app = require('./app');
+const utils = require('./utils/misc');
+const logger = require('./utils/logger');
 const autoCancelation = require('./automation/expiredReservations');
 const autoOpening = require('./automation/gameOpening');
-const dal = require('./dal/dal');
-const events = require('./utils/notifications');
-const logger = require('./utils/logger');
-const utils = require('./utils/misc');
 
-const app = express();
 const config = utils.getConfig();
 const log = logger.create('SERVER');
 
-process.on('unhandledRejection', function(reason){
-  log.error(`unhandledRejection at promise with reason: ${reason.stack}`);
-  events.emit('system.unhandledRejection');
-});
-
-process.on('uncaughtException', function(error) {
-  log.error(`uncaughtException: ${error && error.stack || error}`);
-  events.emit('system.uncaughtException');
-});
-
-app.use(bodyParser.json()); // support json encoded bodies
-app.use(bodyParser.urlencoded({ extended: true })); // support encoded bodies
-app.use(cookieParser());
-app.enable('trust proxy');
-
-app.use((req, res, next) => {
-  const start = Date.now();
-  req.id = uuid.v4().slice(0, 8);
-  req.userId = auth.decode(req.cookies.auth);
-  req.log = logger.create(undefined, req);
-  req.dal = dal;
-
-  req.log.info(`${req.ip} ${req.userId || 'noauth'} request: ${req.path}`);
-  req.on('end', () => {
-    const time = Date.now() - start;
-    req.log.info(`response [${res.statusCode}] ${time}ms`);
-  });
-  next();
-});
-
-apiRoutes.init(app);
-
-app.get('*', function (req, res) {
-  res.status(404).send('Not Found\n');
-});
-
 app.listen(config.server.port, config.server.host);
 log.info(`listen on: ${config.server.host}:${config.server.port}`);
-
 
 // run tasks every minute
 autoCancelation.act();
@@ -64,4 +18,3 @@ setInterval(() => {
   autoCancelation.act();
   autoOpening.act();
 }, 60000);
-
